rename folder test/ethereum
