+++ test/ethereum/token/tokenbytecode.json
